pyproject.toml
[project]
name = "alessia"
version = "1.0.0"
dependencies = [
    "python-dotenv==1.0.0",
    "discord.py==2.2.2",
    "aiohttp==3.8.4",
    "sentry-sdk==1.17.0",
    "rapidfuzz==2.13.7",
    "coloredlogs==15.0.1",
    "psycopg[binary]==3.1.8",
    "SQLAlchemy==2.0.7",
    "pydantic==1.10.6",
]

[project.urls]
repository = "https://github.com/letsbuilda/alessia"
documentation = "https://docs.letsbuilda.dev/alessia/"

[project.optional-dependencies]
dev = [
    "black",
    "isort",
    "pylint",
]
tests = [
    "pytest",
]
docs = [
    "sphinx",
    "furo",
    "sphinx-autoapi",
    "toml",
    "releases",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
target-version = ["py310"]
line-length = 120

[tool.isort]
profile = "black"

[tool.pylint.format]
max-line-length = 120
